Added 'chatroom' to app
"Chatroom" functionality to allow two users to write a story together and then once they are both done the story will be put into the database. The chatroom will use a global class to store the info and there is currently only one chatroom that can be used. Will need to add multiple chatrooms, a lobby, and error checking.